+++ src/components/staff/canditadeMiniCard/candidateMiniCard.scss
