+++ src/components/staff/canditadeMiniCard/candidateMiniCard.tsx
