--- src/components/table/styleTable.scss
@@ -1,24 +1,0 @@
-div {
-  padding: 0.5rem;
-}
-
-table {
-  border-spacing: 0;
-  border: 1px solid black;
-}
-
-td :last-child {
-  border-bottom: 0;
-}
-
-th,
-td {
-  margin: 0;
-  padding: 0.5rem;
-  border-bottom: 1px solid black;
-  border-right: 1px solid black;
-}
-
-td:last-child {
-  border-right: 0;
-}
